ai-server-deploy/requirements.txt
flask
flask-cors
google-generativeai
pdf2image
opencv-python-headless
numpy
pillow
gunicorn
